refactor BigQuery component into multiple components
