+++ src/components/BigQueryExplorer/AIGenerator.tsx
@@ -1,0 +1,188 @@
+import { useEffect, useMemo, useState } from "react";
+import Button from "../Button";
+
+type Props = {
+  nlPrompt: string;
+  setNlPrompt: (v: string) => void;
+  model: string;
+  setModel: (v: string) => void;
+  apiKey: string;
+  setApiKey: (v: string) => void;
+  generating: boolean;
+  onGenerate: () => void;
+  genError: string | null;
+};
+
+const AIGenerator = ({
+  nlPrompt,
+  setNlPrompt,
+  model,
+  setModel,
+  apiKey,
+  setApiKey,
+  generating,
+  onGenerate,
+  genError,
+}: Props) => {
+  const [validatingModel, setValidatingModel] = useState(false);
+  const [modelValid, setModelValid] = useState<boolean | null>(null);
+  const [modelValidationMsg, setModelValidationMsg] = useState<string>("");
+
+  const usingBuiltInAppKey = useMemo(() => {
+    return !apiKey && !!(process.env.REACT_APP_OPENROUTER_API_KEY || process.env.REACT_APP_OPENROUTER_KEY);
+  }, [apiKey]);
+
+  // Debounced model validation against OpenRouter API
+  useEffect(() => {
+    let active = true;
+    if (!model?.trim()) {
+      setModelValid(null);
+      setModelValidationMsg("");
+      return;
+    }
+    const trimmed = model.trim();
+    // If using the app key (built-in), only allow :free models
+    if (usingBuiltInAppKey) {
+      if (!trimmed.endsWith(":free")) {
+        setModelValid(false);
+        setModelValidationMsg("Only free models are supported with the built-in key (suffix :free)");
+        return;
+      }
+    }
+    setValidatingModel(true);
+    setModelValidationMsg("");
+    const slug = trimmed.split(":")[0];
+    const timer = setTimeout(async () => {
+      try {
+        const usedKey = (apiKey || process.env.REACT_APP_OPENROUTER_API_KEY || process.env.REACT_APP_OPENROUTER_KEY || "").trim();
+        if (!usedKey) {
+          setModelValid(null);
+          setModelValidationMsg("");
+          return;
+        }
+        const res = await fetch(`https://openrouter.ai/api/v1/models/${slug}/endpoints`, {
+          headers: {
+            Authorization: `Bearer ${usedKey}`,
+          },
+        });
+        if (!active) return;
+        if (!res.ok) {
+          setModelValid(false);
+          setModelValidationMsg(`Model not found or unauthorized (${res.status})`);
+        } else {
+          const data = await res.json().catch(() => null);
+          const ok = data && (Array.isArray(data) ? data.length > 0 : true);
+          setModelValid(!!ok);
+          if (!ok) setModelValidationMsg("No endpoints available for this model");
+        }
+      } catch (e: any) {
+        if (!active) return;
+        setModelValid(false);
+        setModelValidationMsg("Failed to validate model");
+      } finally {
+        if (active) setValidatingModel(false);
+      }
+    }, 500);
+    return () => {
+      active = false;
+      clearTimeout(timer);
+    };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [apiKey, model, usingBuiltInAppKey]);
+
+  return (
+    <div className="bg-white rounded-xl shadow overflow-hidden mb-6">
+      <div className="p-4">
+        <label className="block text-sm text-gray-600 mb-2">Describe your query</label>
+        <textarea
+          value={nlPrompt}
+          onChange={(e) => setNlPrompt(e.target.value)}
+          placeholder="e.g., List top 10 contracts by verification date"
+          rows={4}
+          className="w-full border border-gray-200 rounded-md px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-ceruleanBlue-300 resize-y"
+        />
+
+        <div className="mt-3 flex flex-col sm:flex-row gap-2 md:gap-3 items-stretch">
+          <div className="flex items-center gap-2 flex-1">
+            <label className="text-sm text-gray-600">Model</label>
+            <a
+              href="https://openrouter.ai/models?order=newest&q=%3Afree"
+              target="_blank"
+              rel="noopener noreferrer"
+              title="Find more free models"
+              aria-label="Find more free models"
+              className="text-gray-400 hover:text-ceruleanBlue-600 transition-colors"
+            >
+              <svg
+                className="h-4 w-4"
+                viewBox="0 0 24 24"
+                fill="none"
+                stroke="currentColor"
+                strokeWidth="2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                aria-hidden="true"
+              >
+                <rect x="3" y="7" width="18" height="10" rx="5" />
+                <circle cx="9" cy="12" r="3" fill="currentColor" stroke="none" />
+              </svg>
+            </a>
+            <div className="w-full">
+              <input
+                value={model}
+                onChange={(e) => setModel(e.target.value)}
+                placeholder="author/slug:free (e.g., meta-llama/llama-3.1-8b-instruct:free)"
+                className="border border-gray-200 rounded-md px-2 py-2 text-sm w-full"
+              />
+            </div>
+            {validatingModel && (
+              <svg className="animate-spin h-4 w-4 text-gray-500" viewBox="0 0 24 24" fill="none">
+                <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
+                <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8v4a4 4 0 00-4 4H4z"></path>
+              </svg>
+            )}
+            {modelValid === true && !validatingModel && (
+              <span className="text-green-600 text-xs">Valid</span>
+            )}
+            {modelValid === false && !validatingModel && (
+              <span className="text-red-600 text-xs" title={modelValidationMsg}>Invalid</span>
+            )}
+          </div>
+
+          <div className="flex-1 flex items-center gap-2 sm:justify-end">
+            <label className="text-sm text-gray-600 shrink-0">OpenRouter key</label>
+            <input
+              type="password"
+              value={apiKey}
+              onChange={(e) => setApiKey(e.target.value)}
+              placeholder="sk-or-..."
+              className="border border-gray-200 rounded-md px-2 py-2 text-sm w-full sm:w-72"
+              autoComplete="off"
+            />
+          </div>
+        </div>
+
+        <div className="mt-3 flex items-center gap-3">
+          <Button onClick={onGenerate} className="uppercase" disabled={generating || modelValid !== true}>
+            {generating ? "Generating…" : "Generate SQL"}
+          </Button>
+          <div className="text-xs text-gray-500">
+            {apiKey ? "Using your OpenRouter key: all models supported." : "Using built-in key: only free models supported"}
+          </div>
+        </div>
+      </div>
+      {!apiKey && !(process.env.REACT_APP_OPENROUTER_API_KEY || process.env.REACT_APP_OPENROUTER_KEY) && (
+        <div className="px-4 py-3 bg-yellow-50 text-yellow-800 border-t border-yellow-200 text-sm">
+          Enter your OpenRouter API key above to enable all models, or configure REACT_APP_OPENROUTER_API_KEY to use free models without entering a key.
+        </div>
+      )}
+      {genError && (
+        <div className="px-4 py-3 bg-red-50 text-red-700 border-t border-red-200 text-sm">
+          {genError}
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default AIGenerator;
