+++ src/components/BigQueryExplorer/constants.ts
@@ -1,0 +1,82 @@
+export const DEFAULT_SQL = `-- Example: Latest verified contract's chain and address
+SELECT chain_id, address
+FROM sourcify_staging.public_verified_contracts vc
+JOIN sourcify_staging.public_contract_deployments cd ON vc.deployment_id = cd.id
+ORDER BY vc.created_at DESC
+LIMIT 1;`;
+
+// Minimal BigQuery schema context for better SQL generation
+export const SCHEMA_DOC = `
+You are writing SQL for BigQuery (Standard SQL) against dataset sourcify_staging.
+All tables are prefixed with public_. Only use these tables/columns and relationships:
+
+- public_code (code_hash BYTEA PRIMARY KEY, code_hash_keccak BYTEA, code BYTEA)
+
+- public_contracts (
+  id UUID PRIMARY KEY,
+  creation_code_hash BYTEA REFERENCES public_code(code_hash),
+  runtime_code_hash BYTEA REFERENCES public_code(code_hash)
+)
+
+- public_compiled_contracts (
+  id UUID PRIMARY KEY,
+  compiler TEXT, version TEXT, language TEXT,
+  name TEXT, fully_qualified_name TEXT,
+  compiler_settings JSONB, compilation_artifacts JSONB,
+  creation_code_hash BYTEA REFERENCES public_code(code_hash),
+  creation_code_artifacts JSONB,
+  runtime_code_hash BYTEA REFERENCES public_code(code_hash),
+  runtime_code_artifacts JSONB
+)
+
+- public_compiled_contracts_sources (
+  id UUID PRIMARY KEY,
+  compilation_id UUID REFERENCES public_compiled_contracts(id),
+  source_hash BYTEA REFERENCES public_sources(source_hash),
+  path TEXT
+)
+
+- public_sources (
+  source_hash BYTEA PRIMARY KEY,
+  source_hash_keccak BYTEA,
+  content TEXT
+)
+
+- public_contract_deployments (
+  id UUID PRIMARY KEY,
+  chain_id BIGINT,
+  address BYTEA,
+  transaction_hash BYTEA,
+  block_number NUMERIC,
+  transaction_index NUMERIC,
+  deployer BYTEA,
+  contract_id UUID REFERENCES public_contracts(id)
+)
+
+- public_verified_contracts (
+  id BIGINT PRIMARY KEY,
+  deployment_id UUID REFERENCES public_contract_deployments(id),
+  compilation_id UUID REFERENCES public_compiled_contracts(id),
+  creation_match BOOLEAN, creation_values JSONB, creation_transformations JSONB, creation_metadata_match BOOLEAN,
+  runtime_match BOOLEAN, runtime_values JSONB, runtime_transformations JSONB, runtime_metadata_match BOOLEAN,
+  created_at TIMESTAMP, updated_at TIMESTAMP
+)
+
+Foreign keys:
+- compiled_contracts.creation_code_hash -> code.code_hash
+- compiled_contracts.runtime_code_hash -> code.code_hash
+- contracts.creation_code_hash -> code.code_hash
+- contracts.runtime_code_hash -> code.code_hash
+- compiled_contracts_sources.compilation_id -> compiled_contracts.id
+- compiled_contracts_sources.source_hash -> sources.source_hash
+- contract_deployments.contract_id -> contracts.id
+- verified_contracts.deployment_id -> contract_deployments.id
+- verified_contracts.compilation_id -> compiled_contracts.id
+
+Conventions:
+- Addresses and hashes are stored as BYTEA; for display, hex-encode as needed.
+- Only in FROM and JOIN clauses, use full table names (e.g., sourcify_staging.public_contracts).
+- Prefer using verified_contracts joined with contract_deployments for chain_id/address queries.
+- Unless specified, include an explicit LIMIT 10.
+`;
+
